+++ src/components/RestaurantMenu.jsx
@@ -1,24 +1,11 @@
-import { useEffect, useState } from "react";
-import ShimmerUi from "./ShimmerUi";
 import { useParams } from "react-router-dom";
-import { MENU_API } from "../utils/constants";
+import useRestaurantMenu from "../utils/useRestaurantMenu";
+import ShimmerUi from "./ShimmerUi";
 
 const RestaurantMenu = () => {
-  const [resInfo, setResInfo] = useState(null);
-
-  const { resId } = useParams();
-
-  useEffect(() => {
-    fetchMenu();
-  }, []);
-
-  const fetchMenu = async () => {
-    const data = await fetch(MENU_API + resId);
-
-    const json = await data.json();
-    // console.log(json);
-    setResInfo(json?.data);
-  };
+  const {resId} = useParams()
+  
+  const resInfo = useRestaurantMenu(resId)
 
   if (resInfo === null) return <ShimmerUi />;
